Moved the logic for resolving logical expressions into abstract_value
This will allow for more precise abstractions to do better than a niave
approach (where we can only resolve the expressions when we know exactly
what they are).

Use simplify expr to evaluate equals

We simplify the two sides of it recursively then pass the whole
expression (with constants replaced) to simplify_expr and see if we get
a result back.